赛季排行弹窗对接json
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@646872 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/pages/team/team_training/team_new/controller.dart
@@ -526,8 +526,8 @@ class TeamController extends GetxController with GetTickerProviderStateMixin {
       //   }
       // }
       if (CacheApi.cupDefineList[i].substituteSum > myTeamEntity.benchCount) {
-        List<double> cupNum = CacheApi.cupDefineList[i].cupNum;
-        double begin = cupNum.first;
+        List<int> cupNum = CacheApi.cupDefineList[i].cupNum;
+        int begin = cupNum.first;
         return begin.ceil();
       }
     }
